refactor: item ページ描画を ItemPage コンポーネントに抽出（/comment ルート共有のため） (#164)
